physical-ai-textbook/api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
asyncpg>=0.29.0
qdrant-client>=1.8.0
openai>=1.3.5
aiofiles>=23.2.1
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.25.0
pydantic-settings>=2.1.0